+++ Hazard-Renderer/src/Backend/OpenGL/OpenGLFramebuffer.h
@@ -20,7 +20,7 @@ namespace HazardRenderer::OpenGL
 
 		virtual uint32_t GetWidth() const override { return m_Specs.Width; };
 		virtual uint32_t GetHeight() const override { return m_Specs.Height; };
-		size_t GetColorAttachmentCount() override { return m_ColorAttachments.size(); }
+		uint64_t GetColorAttachmentCount() override { return m_ColorAttachments.size(); }
 
 		virtual Ref<Image2D> GetImage(uint32_t index = 0) const override { return m_ColorImages[index].As<Image2D>(); };
 		virtual Ref<Image2D> GetDepthImage() const override { return m_DepthImage; }
