+++ Hazard-Renderer/src/Backend/Vulkan/VulkanFramebuffer.h
@@ -25,7 +25,7 @@ namespace HazardRenderer::Vulkan
 
 		uint32_t GetWidth() const override { return m_Specs.Width; };
 		uint32_t GetHeight() const override { return m_Specs.Height; };
-		size_t GetColorAttachmentCount() override { return m_ColorAttachments.size(); }
+		uint64_t GetColorAttachmentCount() override { return m_ColorAttachments.size(); }
 
 		Ref<Image2D> GetImage(uint32_t index = 0) const override { return m_ColorAttachments[index]; };
 		Ref<Image2D> GetDepthImage() const override { return m_DepthAttachmentImage; }
