+++ Hazard-Utility/src/Batch.h
@@ -7,7 +7,7 @@ class Batch
 {
 public:
 	Batch() = default;
-	Batch(size_t size)
+	Batch(uint64_t size)
 	{
 		Resize(size);
 	};
@@ -21,13 +21,13 @@ public:
 		m_DataBuffer.Write(&data, sizeof(Type), m_WriteBufferOffset);
 		m_WriteBufferOffset += sizeof(Type);
 	};
-	inline void AddIndices(size_t count) { m_IndexCount += count; }
+	inline void AddIndices(uint64_t count) { m_IndexCount += count; }
 	void Reset() 
 	{
 		m_WriteBufferOffset = 0;
 		m_IndexCount = 0;
 	};
-	void Resize(size_t size)
+	void Resize(uint64_t size)
 	{
 		m_DataBuffer.Release();
 		m_DataBuffer.Allocate(size * sizeof(Type));
@@ -36,13 +36,13 @@ public:
 	}
 
 	const void* GetData() const { return m_DataBuffer.Data; }
-	size_t GetCount() { return m_WriteBufferOffset / sizeof(Type); }
-	size_t GetSize() { return m_DataBuffer.Size; }
-	size_t GetDataSize() { return m_WriteBufferOffset; }
-	size_t GetIndexCount() { return m_IndexCount; }
+	uint64_t GetCount() { return m_WriteBufferOffset / sizeof(Type); }
+	uint64_t GetSize() { return m_DataBuffer.Size; }
+	uint64_t GetDataSize() { return m_WriteBufferOffset; }
+	uint64_t GetIndexCount() { return m_IndexCount; }
 	operator bool() const { return m_IndexCount != 0; }
 private:
-	size_t m_IndexCount = 0;
-	size_t m_WriteBufferOffset = 0;
+	uint64_t m_IndexCount = 0;
+	uint64_t m_WriteBufferOffset = 0;
 	Buffer m_DataBuffer;
 };
